changed the hide_unavailable_items logic to work on DB directly
Changed filtering items depending on if they have stock or not through the option (hide unavailable items) in POS profile to be applied as a condition while fetching the data from the database instead of looping through the db result and then filtering.

this results in a better performance when the (Hide Unavailable Items) is enabled.